✨ Feature - Article, Question Search (#39)
* feat: search in article

* fix: wrong sql

* feat: search in question list

* fix: divide query cause of pageable with fetch join

* fix: remove unnecessary variables, import

+++ server-api/src/main/java/org/dongguk/dscd/wooahan/api/question/dto/response/ReadQuestionListDto.java
@@ -26,7 +26,7 @@ public record ReadQuestionListDto (
             @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
             LocalDateTime createdAt,
             @JsonProperty("answer_count")
-            Integer answerCount,
+            Long answerCount,
             @JsonProperty("nickname")
             String nickname,
             @JsonProperty("creator_id")
